Backend/requirements.txt
fastapi>=0.104.1
pydantic>=2.4.2
email-validator>=2.1.0
alembic>=1.12.1
SQLAlchemy>=2.0.23
psycopg2-binary>=2.9.9
Jinja2>=3.1.2
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
pytest>=7.4.3
httpx>=0.26.0
openai>=1.3.3
python-multipart>=0.0.6
PyYAML>=6.0.1
python-dotenv>=1.0.0
uvicorn>=0.23.2
pyngrok>=7.0.0
aiosmtpd>=1.4.4.post2
ipython>=8.17.2
pytest-asyncio>=0.21.1
black>=23.11.0
ruff>=0.1.5
pytest-cov>=4.1.0
python-docx==0.8.11
celery==5.3.4
redis==5.0.1
pika==1.3.2
tenacity==8.2.3
sentry-sdk==1.37.1
pymongo==4.6.1
google-generativeai==0.3.1
pinecone-client==2.2.4
sentence-transformers==2.2.2
transformers==4.37.2
safetensors==0.4.1
torch==2.1.2
requests==2.31.0
re>=2.2.1 